backend/requirements.txt
fastapi
uvicorn[standard]
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.10
openai>=1.12.0
youtube-transcript-api>=0.6.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.0
